+++ packages/rx_bloc_list/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [3.2.0-dev.1] January 30, 2022
 - Added `Identifiable` interface for handling inter-feature communication
 - Added `List<T extends Identifiable>` extensions: `ids`, `containsIdentifiable`, `removeIdentifiable`, `mergeWith`
-- Added `Stream<T extends Identifiable>` extensions: `mapCreatedWithLatestFrom`, `mapUpdatedWithLatestFrom`, `mapDeletedWithLatestFrom`
+- Added `Stream<T extends Identifiable>` extensions: `withLatestFromIdentifiableList`
 
 ## [3.1.0] - September 24, 2021
 * Support Flutter `2.5`